test(csa-client): ETXTBSY 対策の write/spawn 直列化を tests/common に共通化し姉妹テストにも適用
engine_info_cache / session_events_integration にも「script を書いた直後に
UsiEngine::spawn を直接呼ぶ」同一パターンが残っており、同種の ETXTBSY flake
が顕在化し得るとのレビュー指摘への対応。

FORK_WRITE_LOCK / write_mock_script / spawn_engine を tests/common/mod.rs に
切り出し、3 テストファイルすべての script 書き込みと spawn を同一 lock で
直列化する。lock が handshake 完了まで保持される点の注意書きも doc comment
に追記 (レビュー補足指摘)。
